+++ src/Models/Test/ModelInstantiatorTest.php
@@ -4,6 +4,8 @@ namespace PDGA\DataObjects\Models\Test;
 
 use \DateTime;
 
+use PDGA\DataObjects\Models\Test\ModelInstantiatorTestObject;
+use PDGA\DataObjects\Models\Test\PrivacyProtectedTestDataObject;
 use PHPUnit\Framework\TestCase;
 
 use PDGA\Exception\InvalidRelationshipDataException;
@@ -14,7 +16,6 @@ use PDGA\DataObjects\Models\ModelInstantiator;
 use PDGA\DataObjects\Models\ReflectionContainer;
 use PDGA\DataObjects\Models\Test\Member;
 use PDGA\DataObjects\Models\Test\ModelInstantiatorTestDBModel;
-use PDGA\DataObjects\Models\Test\ModelInstantiatorTestObject;
 use PDGA\DataObjects\Models\Test\PhoneNumber;
 
 class ModelInstantiatorTest extends TestCase
@@ -316,6 +317,8 @@ class ModelInstantiatorTest extends TestCase
 
         // The output array should reflect the Data Object properties.
         // Note that the order of the array keys matters and must match the class definition.
+        $result = $this->model_instantiator->dataObjectToArray($data_object);
+
         $this->assertSame(
             [
                 'pdgaNumber'   => 4297,
@@ -326,7 +329,117 @@ class ModelInstantiatorTest extends TestCase
                 'birthDate'    => '2020-01-01T00:00:00+00:00',
                 'testProperty' => true,
             ],
-            $this->model_instantiator->dataObjectToArray($data_object)
+            $result
+        );
+    }
+
+    public function testDataObjectToArrayWithPrivacyProtectedDataObject(): void
+    {
+        // Create an input Privacy Protected Data Object instance.
+        $data_object = $this->getPrivacyProtectedTestDataObject();
+
+        // Assert that the privacy protected properties are set prior to array conversion.
+        $this->assertPrivacyProtectedPropertiesAreSet($data_object);
+
+        // The output array should reflect the Data Object properties.
+        // The values for the privacy protected fields should be removed.
+        $result = $this->model_instantiator->dataObjectToArray($data_object);
+
+        $this->assertSame(
+            [
+                'pdgaNumber'   => 4297,
+                'firstName'    => 'Ken',
+                'lastName'     => 'Climo',
+                'testProperty' => true,
+            ],
+            $result
+        );
+
+        foreach (PrivacyProtectedTestDataObject::PRIVACY_PROTECTED_PROPERTIES as $property) {
+            $this->assertFalse(isset($result[$property]));
+        }
+    }
+
+    public function testDataObjectToArrayWithPrivacyProtectedDataObjectButAlsoOverridden(): void
+    {
+        // Create an input Data Object instance.
+        $data_object               = new ModelInstantiatorTestObject();
+        $data_object->firstName    = 'Ken';
+        $data_object->lastName     = 'Climo';
+        $data_object->pdgaNumber   = 4297;
+        $data_object->email        = 'champ@pdga.com';
+        $data_object->privacy      = true;
+        $data_object->testProperty = true;
+        $data_object->birthDate    = new DateTime('2020-01-01');
+
+        // The output array should reflect the Data Object properties.
+        // Note that the order of the array keys matters and must match the class definition.
+        $cleanse_privacy = false;
+        $result = $this->model_instantiator->dataObjectToArray($data_object, $cleanse_privacy);
+
+        $this->assertSame(
+            [
+                'pdgaNumber'   => 4297,
+                'firstName'    => 'Ken',
+                'lastName'     => 'Climo',
+                'email'        => 'champ@pdga.com',
+                'privacy'      => true,
+                'birthDate'    => '2020-01-01T00:00:00+00:00',
+                'testProperty' => true,
+            ],
+            $result
+        );
+    }
+
+    public function testNestedDataObjectToArrayWithPrivacyProtectedDataObject(): void
+    {
+        // Create a related Privacy Protected Data Object instance.
+        $fake_has_one_data_object = $this->getPrivacyProtectedTestDataObject();
+        $nullable_fake_has_one_data_object = $this->getPrivacyProtectedTestDataObject();
+        $fake_has_many_data_object = [$this->getPrivacyProtectedTestDataObject()];
+
+        // Create the primary data object and set relationship values.
+        $data_object = $this->getPrivacyProtectedTestDataObject();
+        $data_object->fakeHasOneRelation = $fake_has_one_data_object;
+        $data_object->nullableFakeHasOneRelation = $nullable_fake_has_one_data_object;
+        $data_object->fakeHasManyRelation = $fake_has_many_data_object;
+
+        // Assert that the privacy protected properties are set prior to array conversion.
+        $this->assertPrivacyProtectedPropertiesAreSet($data_object);
+        $this->assertPrivacyProtectedPropertiesAreSet($fake_has_one_data_object);
+        $this->assertPrivacyProtectedPropertiesAreSet($nullable_fake_has_one_data_object);
+        $this->assertPrivacyProtectedPropertiesAreSet($fake_has_many_data_object[0]);
+
+        // The output array should reflect the Data Object properties.
+        // The values for the privacy protected fields should be removed.
+        $result = $this->model_instantiator->dataObjectToArray($data_object);
+
+        $this->assertSame(
+            [
+                'pdgaNumber'   => 4297,
+                'firstName'    => 'Ken',
+                'lastName'     => 'Climo',
+                'testProperty' => true,
+                'fakeHasOneRelation' => [
+                    'pdgaNumber'   => 4297,
+                    'firstName'    => 'Ken',
+                    'lastName'     => 'Climo',
+                    'testProperty' => true,
+                ],
+                'nullableFakeHasOneRelation' => [
+                    'pdgaNumber'   => 4297,
+                    'firstName'    => 'Ken',
+                    'lastName'     => 'Climo',
+                    'testProperty' => true,
+                ],
+                'fakeHasManyRelation' => [[
+                    'pdgaNumber'   => 4297,
+                    'firstName'    => 'Ken',
+                    'lastName'     => 'Climo',
+                    'testProperty' => true,
+                ]],
+            ],
+            $result
         );
     }
 
@@ -511,6 +624,31 @@ class ModelInstantiatorTest extends TestCase
             $this->assertTrue(false);
         } catch (InvalidRelationshipDataException $e) {
             $this->assertEquals('FakeHasOneRelation relationship must not be null.', $e->getMessage());
+        }
+    }
+
+    /**
+     * @return PrivacyProtectedTestDataObject
+     */
+    private function getPrivacyProtectedTestDataObject(): PrivacyProtectedTestDataObject
+    {
+        // Create an input Privacy Protected Data Object instance.
+        $data_object = new PrivacyProtectedTestDataObject();
+        $data_object->firstName = 'Ken';
+        $data_object->lastName = 'Climo';
+        $data_object->pdgaNumber = 4297;
+        $data_object->email = 'champ@pdga.com';
+        $data_object->privacy = true;
+        $data_object->testProperty = true;
+        $data_object->birthDate = new DateTime('2020-01-01');
+
+        return $data_object;
+    }
+
+    private function assertPrivacyProtectedPropertiesAreSet(PrivacyProtectedTestDataObject $data_object): void
+    {
+        foreach (PrivacyProtectedTestDataObject::PRIVACY_PROTECTED_PROPERTIES as $property) {
+            $this->assertTrue(isset($data_object->$property));
         }
     }
 }